EG5-130: Refactor del codigo - Movido logica del Controller a Service.
