+++ src/main/java/com/eoi/grupo5/services/foro/HiloLookupService.java
@@ -1,0 +1,28 @@
+package com.eoi.grupo5.services.foro;
+
+import com.eoi.grupo5.entities.foro.EntidadHilo;
+import com.eoi.grupo5.repositories.foro.HiloRepository;
+import org.springframework.stereotype.Service;
+
+/**
+ * Decision por tener una tercera service, Spring no me dejaba tener un bucle de inyecciones entre Mensaje y Hilo service.
+ * Entonces he tenido que crear este service extra para los metodos que hacen falta en el MensajeService.
+ */
+
+@Service
+public class HiloLookupService {
+    private final HiloRepository hiloRepository;
+
+    public HiloLookupService(HiloRepository hiloRepository) {
+        this.hiloRepository = hiloRepository;
+    }
+
+    public EntidadHilo findById(Long id) {
+        return hiloRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Hilo no encontrado"));
+    }
+
+    public void saveHilo(EntidadHilo hilo) {
+        hiloRepository.save(hilo);
+    }
+}
