+++ src/main/resources/templates/foro/hilo.html
@@ -31,7 +31,7 @@
 
 <body>
     <!-- Incluir navbar -->
-    <div th:replace="plantillas/navbar :: navbar"></div>
+    <div th:replace="~{plantillas/navbar :: navbar}"></div>
 
     <!-- Main Content Area -->
     <div class="container mt-3">
@@ -66,11 +66,6 @@
                             May 27, 2025 3:30 PM
                         </time>
                     </div>
-
-                    <!-- Right side: Me Gusta button -->
-                    <a class="btn btn-sm btn-outline-dark">
-                        <i class="fa-solid fa-heart"></i> Me Gusta!
-                    </a>
                 </div>
             </div>
         </section>
@@ -122,7 +117,7 @@
                         aria-hidden="true">
                         <div class="modal-dialog modal-lg">
                             <div class="modal-content">
-                                <form th:action="@{'/foro/mensaje/editar/' + ${mensaje.id}}" method="post">
+                                <form th:action="@{'/foro/hilo/edit-mensaje/' + ${mensaje.id}}" method="post">
                                     <div class="modal-header justify-content-center">
                                         <h3 class="modal-title">
                                             <strong>Editar Mensaje</strong> <i class="fa-solid fa-pen-to-square"></i>
@@ -160,9 +155,8 @@
                                         deshacer.</p>
                                 </div>
                                 <div class="modal-footer">
-                                    <form th:action="@{'/foro/mensaje/eliminar/' + ${mensaje.id}}" method="post">
-                                        <input type="hidden" th:name="${_csrf.parameterName}"
-                                            th:value="${_csrf.token}" />
+                                    <form th:action="@{'/foro/hilo/delete-mensaje/' + ${mensaje.id}}" method="post">
+
                                         <button type="submit" class="btn btn-danger">Borrar</button>
                                         <button type="button" class="btn btn-secondary"
                                             data-bs-dismiss="modal">Cancelar</button>
@@ -178,7 +172,7 @@
         <!-- Reply Form -->
         <section class="reply-form">
             <h2 class="h4 mb-2">Responde a este hilo</h2>
-            <form method="post" th:action="@{/foro/hilo/{id}(id=${hilo.id})}" th:object="${nuevoMensaje}">
+            <form method="post" th:action="@{/foro/hilo/{id}/nuevo-mensaje(id=${hilo.id})}" th:object="${nuevoMensaje}">
                 <div class="mb-2">
                     <textarea name="contenido" th:field="*{contenido}" rows="5" class="form-control"
                         placeholder="Escribe tu respuesta aquí..." required></textarea>
@@ -200,7 +194,7 @@
                 <div class="modal-body">
                     ¿Estás seguro de que deseas borrar este hilo? Esta acción no se puede deshacer.
                 </div>
-                <form th:action="@{'/foro/eliminar/' + ${hilo.id}}" method="post">
+                <form th:action="@{'/foro/delete-hilo/' + ${hilo.id}}" method="post">
                     <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}" />
                     <div class="modal-footer">
                         <button type="submit" class="btn btn-danger">Borrar</button>
@@ -215,7 +209,7 @@
     <div class="modal fade" id="editarHiloModal" tabindex="-1" aria-hidden="true">
         <div class="modal-dialog modal-lg">
             <div class="modal-content">
-                <form th:action="@{'/foro/editar/' + ${hilo.id}}" method="post">
+                <form th:action="@{'/foro/edit-hilo/' + ${hilo.id}}" method="post">
                     <div class="modal-header justify-content-center">
                         <h3 class="modal-title">
                             <strong>Editar Hilo</strong> <i class="fa-solid fa-pen-to-square"></i>
@@ -243,7 +237,7 @@
 
 
     <!-- Incluir footer -->
-    <div th:replace="plantillas/footer :: footer"></div>
+    <div th:replace="~{plantillas/footer :: footer}"></div>
 
 </body>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"
