+++ src/main/resources/templates/foro/main.html
@@ -60,7 +60,7 @@
     <div class="modal fade" id="nuevoHiloModal" tabindex="-1" aria-hidden="true">
       <div class="modal-dialog modal-lg">
         <div class="modal-content">
-          <form th:action="@{/foro/crear}" th:object="${nuevoHilo}" method="post">
+          <form th:action="@{/foro}" th:object="${nuevoHilo}" method="post">
             <div class="modal-header justify-content-center">
               <h3 class="modal-title"><strong>Nuevo Hilo!</strong> <i class="fa-solid fa-pencil"></i></h3>
             </div>
